Simplify mesh management
