+++ src/main/kotlin/org/janelia/saalfeldlab/paintera/meshes/managed/GetBlockListFor.kt
@@ -1,0 +1,7 @@
+package org.janelia.saalfeldlab.paintera.meshes.managed
+
+import net.imglib2.Interval
+
+interface GetBlockListFor<Key> {
+    fun getBlocksFor(level: Int, key: Key): Array<Interval>?
+}
